rubocop

+++ test_helpers.rb
@@ -383,7 +383,7 @@ def parse_total_site_energy(out_osw_path)
     return nil
   end
 
-  if !result_osw.has_key?(:steps)
+  if !result_osw.key?(:steps)
     # puts "#{out_osw_path} doesn't have :steps"
     return nil
   end
@@ -393,7 +393,7 @@ def parse_total_site_energy(out_osw_path)
     return nil
   end
 
-  if !result_osw[:steps][0].has_key?(:result)
+  if !result_osw[:steps][0].key?(:result)
     # puts "#{out_osw_path} has steps without :result"
     return nil
   end
@@ -513,7 +513,7 @@ def postprocess_out_osw_and_copy(out_osw, cp_out_osw, compare_eui = true)
         end
       end
     else
-      raise "postprocess_out_osw_and_copy: there should always be one openstudio_results measure!"
+      raise 'postprocess_out_osw_and_copy: there should always be one openstudio_results measure!'
     end
 
     # The fuel cell tests produce out.osw files that are about 800 MB in old
@@ -566,10 +566,10 @@ def sim_test(filename, options = {})
     base_dir = $ModelDir
   end
 
-  if !options[:compare_eui].nil?
-    compare_eui = options[:compare_eui]
-  else
+  if options[:compare_eui].nil?
     compare_eui = true
+  else
+    compare_eui = options[:compare_eui]
   end
 
   # puts "Running sim_test(#{filename})"
@@ -580,17 +580,13 @@ def sim_test(filename, options = {})
 
   # If $DoNotReRunIfSuccess is true, we check if the out_osw already exists
   # and whether it was successful already
-  if $DoNotReRunIfSuccess
-    if File.exist?(cp_out_osw)
-      cp_result_osw = nil
-      File.open(cp_out_osw, 'r') do |f|
-        cp_result_osw = JSON.parse(f.read, symbolize_names: true)
-      end
-      if !cp_result_osw.nil?
-        if cp_result_osw[:completed_status] == 'Success'
-          skip 'Already ran with success'
-        end
-      end
+  if $DoNotReRunIfSuccess && File.exist?(cp_out_osw)
+    cp_result_osw = nil
+    File.open(cp_out_osw, 'r') do |f|
+      cp_result_osw = JSON.parse(f.read, symbolize_names: true)
+    end
+    if !cp_result_osw.nil? && (cp_result_osw[:completed_status] == 'Success')
+      skip 'Already ran with success'
     end
   end
 
@@ -603,7 +599,8 @@ def sim_test(filename, options = {})
   FileUtils.mkdir_p(dir)
 
   ext = File.extname(filename)
-  if (ext == '.osm') || (ext == '.xml')
+  case ext
+  when '.osm', '.xml'
 
     if ext == '.xml'
       new_filename = filename.sub('.xml', '.osm')
@@ -656,7 +653,7 @@ def sim_test(filename, options = {})
       FileUtils.cp(sch_ori_path, sch_target_path)
     end
 
-  elsif ext == '.rb'
+  when '.rb'
 
     # Copy the generic OSW file, needed to add design days in particular when
     # running the measure to generate the OSM, and then of course for the sim
@@ -673,7 +670,7 @@ def sim_test(filename, options = {})
       FileUtils.mv(out_osm, in_osm)
     end
 
-  elsif ext == '.osw'
+  when '.osw'
 
     # make an empty osm
     model = OpenStudio::Model::Model.new
@@ -773,9 +770,10 @@ def autosizing_test(filename, weather_file = nil, model_measures = [], energyplu
     FileUtils.cp($OswFile, in_osw)
 
     ext = File.extname(filename)
-    if ext == '.osm'
+    case ext
+    when '.osm'
       FileUtils.cp(File.join($ModelDir, filename), in_osm)
-    elsif ext == '.rb'
+    when '.rb'
       command = "\"#{$OpenstudioCli}\" \"#{File.join($ModelDir, filename)}\""
       run_command(command, dir, 3600)
 
@@ -1013,8 +1011,8 @@ def autosizing_test(filename, weather_file = nil, model_measures = [], energyplu
 
       # Replace the getter name with known alias, if one exists
       obj_aliases = getter_aliases[os_type]
-      if obj_aliases
-        getter_name = obj_aliases[getter_name] unless obj_aliases[getter_name].nil?
+      if obj_aliases && !obj_aliases[getter_name].nil?
+        getter_name = obj_aliases[getter_name]
       end
 
       # Don't test this getter if it is designated to be skipped
@@ -1022,8 +1020,8 @@ def autosizing_test(filename, weather_file = nil, model_measures = [], energyplu
 
       # Don't test this getter if it is known to be missing from E+ output
       obj_missing_getters = missing_getters[os_type]
-      if obj_missing_getters
-        next if obj_missing_getters.include?(getter_name)
+      if obj_missing_getters&.include?(getter_name)
+        next
       end
 
       # Check if the autosizedFoo method has been implemented for this object
@@ -1250,7 +1248,7 @@ def sql_test(options = {})
     FileUtils.mkdir_p(dir)
     # FileUtils.cp($OswFile, osw)
 
-    # Note: If you give E+ a weather file that doesn't have a leap day
+    # NOTE: If you give E+ a weather file that doesn't have a leap day
     # It won't give you an output that has a leap day...
     # And it doesn't appear to let you know about it either
     if options[:isLeapYear]
@@ -1341,7 +1339,8 @@ def sql_test(options = {})
   # Check run period and annual series
   assert_equal(run_periods.dateTimes.length, 1)
   assert_equal(run_periods.values.length, 1)
-  if options[:type] == 'Full'
+  case options[:type]
+  when 'Full'
     assert(ts.is_initialized, "Timeseries isn't initialized")
     annuals = ts.get
     assert_equal(annuals.dateTimes.length, 1)
@@ -1349,9 +1348,9 @@ def sql_test(options = {})
     assert_equal(run_periods.dateTimes[0], annuals.dateTimes[0])
     assert_equal(run_periods.values[0], annuals.values[0])
     assert(annuals.values[0] == run_periods.values[0])
-  elsif options[:type] == 'Partial'
+  when 'Partial'
     assert(!ts.is_initialized, 'Timeseries is initialized')
-  elsif options[:type] == 'Wrap-around'
+  when 'Wrap-around'
     assert(ts.is_initialized, "Timeseries isn't initialized")
     annuals = ts.get
     assert_equal(run_periods.dateTimes.length, 1)
